kernel: HACK: Introduce fake dotconfig
This patch does creating a fake /proc/config.gz.
It will allow us to specify which keys that we want to "tell"
the system that we have but not actually present in the kernel binary.

It does the job by merging $(objtree)/.config and $(srctree)/.fakeconfig,
and then use the merged config (in this case, $(objtree)/.fakeconfig)
as the content of config_data.gz.

This hacks is disabled by default.
Use KBUILD_FAKECONFIG=1 in order to activate it.

Signed-off-by: Nanda Oktavera <codeharuka.yusa@gmail.com>